pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lfm-physics"
version = "1.2.0"
description = "Lattice Field Medium physics simulation library"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Greg Partin", email = "gpartin@gmail.com" },
]
keywords = ["physics", "simulation", "lattice", "field-theory", "klein-gordon"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Physics",
]
dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
]

[project.optional-dependencies]
gpu = ["cupy-cuda12x>=12.0"]
viz = ["matplotlib>=3.7"]
all = ["cupy-cuda12x>=12.0", "matplotlib>=3.7"]
dev = [
    "pre-commit>=3.7",
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
]
docs = [
    "sphinx>=7.0",
    "furo>=2024.1",
    "sphinx-autodoc-typehints>=1.24",
    "myst-parser>=2.0",
]
benchmark = [
    "pytest-benchmark>=4.0",
]

[project.urls]
Homepage = "https://github.com/gpartin/lfm-physics"
Repository = "https://github.com/gpartin/lfm-physics"
Issues = "https://github.com/gpartin/lfm-physics/issues"

[tool.hatch.build.targets.wheel]
packages = ["lfm"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.lint.per-file-ignores]
# CUDA C kernel strings — lines inherently long
"lfm/core/backends/kernel_source.py" = ["E501"]

[tool.coverage.run]
source = ["lfm"]
omit = [
    "lfm/viz/*",
    "lfm/core/backends/kernel_source.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__",
]
show_missing = true
